refactor(iota-data-ingestion-core): remove `CheckpointReaderV1` (#11185)
# Description of change

This PR as part of deprecation effort removes the `CheckpointReaderV1`.
With the pruning enabled on the fullnodes more of the custom indexers
will need to use the `CheckpointReaderV2`to use hybrid historical
storage to sync from genesis, making the `CheckpointReaderV1` obsolete.

## Links to any relevant issues

fixes #10967 

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)

- Refactored the tests to use the `CheckpointReaderV2` as default
- Public re-exports (`ReaderOptions`, etc.) remain unchanged to avoid
breaking imports for operators already using `CheckpointReaderV2`.
- Ran `iota-data-ingestion-core` tests

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [ ] Verification of API backward compatibility.

> [!NOTE]
> This patch does not affect the normal operation of the indexer, thus
no additional test were conducted.

### Release Notes
- [x] Indexer: ⚠️`iota-data-ingestion-core`, the
`IndexerExecutor::run()` is removed as part of `CheckpointReaderV1`
deprecation effort. Users shall migrate to
`IndexerExecutor::run_with_config()` instead which will use the
`CheckpointReaderV2` as default. In this regard `setup_single_workflow`
also had breaking changes, the `remote_store_url` accepts a `RemoteUrl`
instead of a `String`.⚠️
