+++ crates/iota-data-ingestion-core/Cargo.toml
@@ -36,7 +36,6 @@ iota-grpc-client.workspace = true
 iota-grpc-types.workspace = true
 iota-metrics.workspace = true
 iota-protocol-config.workspace = true
-iota-sdk-types.workspace = true
 iota-storage.workspace = true
 iota-types.workspace = true
