+++ crates/iota-data-ingestion-core/src/lib.rs
@@ -50,7 +50,7 @@ pub use executor::{
 use iota_types::full_checkpoint_content::CheckpointData;
 pub use metrics::DataIngestionMetrics;
 pub use progress_store::{FileProgressStore, ProgressStore, ShimProgressStore};
-pub use reader::v1::ReaderOptions;
+pub use reader::ReaderOptions;
 pub use reducer::Reducer;
 pub use util::{create_remote_store_client, create_remote_store_client_with_ops};
 pub use worker_pool::WorkerPool;
