+++ crates/iota-data-ingestion-core/src/reader/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod common;
 pub(crate) mod fetch;
-pub(crate) mod v1;
 pub mod v2;
+
+pub use common::ReaderOptions;
